Apply /simplify findings: helper paths, presence, dedup tests, shared visual helper
- contact_controller: redirect_to contact_path (×2) instead of literal '/contact'
- feedback_mailer: reply_to: @feedback.email.presence — drop the splat-or-empty hash
- feedback_test: drop 'contact is a valid source' (already covered by valid-sources loop)
- visual specs: extract test/visual/helpers.ts for the Anthropic client guard,
  PNG height-crop fallback, and messages.create call. Both email and marketing
  specs now share the subtle 7900px height threshold rather than maintaining it twice.
