+++ test/visual/email-screenshots.spec.ts
@@ -1,12 +1,8 @@
 import { test, expect } from "@playwright/test";
-import Anthropic from "@anthropic-ai/sdk";
 import { readFileSync } from "fs";
 import { join } from "path";
+import { client, cropIfTooTall, reviewScreenshot } from "./helpers";
 
-// Anthropic SDK throws at construction without ANTHROPIC_API_KEY, so defer
-// until we know the key exists. CI doesn't run these specs (per CLAUDE.md);
-// this lets local `bunx playwright test` skip cleanly when the key is unset.
-const client = process.env.ANTHROPIC_API_KEY ? new Anthropic() : null;
 const prompt = readFileSync(join(__dirname, "email-check-prompt.txt"), "utf-8");
 
 const emails = [
@@ -17,46 +13,6 @@ const emails = [
   { mailer: "reminder_mailer", action: "havent_ordered_email" },
   { mailer: "anomaly_mailer", action: "anomaly_report" },
 ];
-
-async function assertEmailLooksGood(screenshot: Buffer, page: any): Promise<string> {
-  // If the full-page screenshot exceeds the 8000px API limit, take a
-  // viewport-height crop instead (checks the top portion of the email).
-  let imageBuffer = screenshot;
-  try {
-    // Quick size check via PNG header — height is at bytes 20-23
-    const height = screenshot.readUInt32BE(20);
-    if (height > 7900) {
-      imageBuffer = await page.screenshot({ fullPage: false });
-    }
-  } catch {
-    // If header read fails, use original
-  }
-
-  const response = await client!.messages.create({
-    model: "claude-haiku-4-5",
-    max_tokens: 256,
-    messages: [
-      {
-        role: "user",
-        content: [
-          {
-            type: "image",
-            source: {
-              type: "base64",
-              media_type: "image/png",
-              data: imageBuffer.toString("base64"),
-            },
-          },
-          { type: "text", text: prompt },
-        ],
-      },
-    ],
-  });
-
-  const result =
-    response.content[0].type === "text" ? response.content[0].text : "";
-  return result.trim().toLowerCase();
-}
 
 for (const { mailer, action } of emails) {
   test(`${mailer}/${action}`, async ({ page }, testInfo) => {
@@ -71,9 +27,8 @@ for (const { mailer, action } of emails) {
       fullPage: true,
     });
 
-    const result = await assertEmailLooksGood(screenshot, page);
-    expect(result, `Visual check failed for ${mailer}/${action}`).toContain(
-      "pass",
-    );
+    const cropped = await cropIfTooTall(screenshot, page);
+    const result = (await reviewScreenshot(cropped, prompt)).trim().toLowerCase();
+    expect(result, `Visual check failed for ${mailer}/${action}`).toContain("pass");
   });
 }
